switch back to handler
all the things we schedule are mainly doing stuff on the main thread.
when using ScheduledThreadPoolExecutor, we need define a runnable and
in there another runnable to run on the main UI thread. This is
annoying and very verbose. The handler way, is not too nice, but less
lines win.